refactor: move out some utils
